refactor(repository): change datatype for similarity

+++ internal/domain/repository/rag/rag.sql
@@ -10,7 +10,7 @@ SELECT
     chunk_index,
     content,
     created_at,
-    1 - (embedding <=> $1) AS similarity
+    (1 - (embedding <=> $1))::float8 AS similarity
 FROM document_chunks
 ORDER BY embedding <=> $1
 LIMIT $2;
